mv pipeline.RegionalCombinedData to combined_datasets.RegionalCombinedData
