+++ libs/pipeline.py
@@ -2,15 +2,10 @@
 Code that is used to help move information around in the pipeline, starting with `Region` which
 represents a geographical area (state, county, metro area, etc).
 """
-from typing import Dict, Any
 from dataclasses import dataclass
 
 import structlog
 import us
-
-from covidactnow.datapublic.common_fields import CommonFields
-from libs.datasets import combined_datasets
-from libs.datasets import timeseries
 
 _log = structlog.get_logger()
 
@@ -53,40 +48,3 @@ class Region:
         if len(self.fips) != 5:
             raise ValueError(f"No state for {self}")
         return Region(fips=self.fips[:2])
-
-
-@dataclass(frozen=True)
-class RegionalCombinedData:
-    """Identifies a geographical area and wraps access to `combined_datasets` of it."""
-
-    region: Region
-
-    latest: Dict[str, Any]
-
-    timeseries: timeseries.TimeseriesDataset
-
-    @staticmethod
-    def from_region(region: Region) -> "RegionalCombinedData":
-
-        us_latest = combined_datasets.load_us_latest_dataset()
-        region_latest = us_latest.get_record_for_fips(region.fips)
-
-        us_timeseries = combined_datasets.load_us_timeseries_dataset()
-        region_timeseries = us_timeseries.get_subset(fips=region.fips)
-
-        return RegionalCombinedData(
-            region=region, latest=region_latest, timeseries=region_timeseries
-        )
-
-    @property
-    def population(self) -> int:
-        """Gets the population for this region."""
-        return self.latest[CommonFields.POPULATION]
-
-    @property  # TODO(tom): Change to cached_property when we're using Python 3.8
-    def display_name(self) -> str:
-        county = self.latest[CommonFields.COUNTY]
-        state = self.latest[CommonFields.STATE]
-        if county:
-            return f"{county}, {state}"
-        return state
