+++ pyseir/ensembles/ensemble_runner.py
@@ -8,6 +8,7 @@ import structlog
 import copy
 from collections import defaultdict
 
+from libs.datasets import combined_datasets
 from libs import pipeline
 from pyseir.inference import model_fitter
 from pyseir.models import seir_model
@@ -30,7 +31,7 @@ compartment_to_capacity_attr_map = {
 class RegionalInput:
     region: pipeline.Region
 
-    _combined_data: pipeline.RegionalCombinedData
+    _combined_data: combined_datasets.RegionalCombinedData
     _mle_fit_model: seir_model.SEIRModel
     _mle_fit_result: Mapping[str, Any]
     _state_mle_fit_model: Optional[seir_model.SEIRModel] = None
